+++ src/Api/idmsvc/.openapi-generator/VERSION
