+++ src/Routes/DetailPage/Components/DetailGeneral/DetailGeneral.tsx
@@ -15,7 +15,7 @@ import {
 } from '@patternfly/react-core';
 import React from 'react';
 import { useState } from 'react';
-import { Domain, ResourcesApiFactory } from '../../../../Api';
+import { Domain, ResourcesApiFactory } from '../../../../Api/idmsvc';
 import useNotification from '../../../../Hooks/useNotification';
 import OutlinedQuestionCircleIcon from '@patternfly/react-icons/dist/esm/icons/outlined-question-circle-icon';
 import PencilAltIcon from '@patternfly/react-icons/dist/esm/icons/pencil-alt-icon';
